template/agents/pyproject.toml
[project]
name = "PROJECT_NAME-agents"
version = "0.1.0"
description = "AI agents powered by Agno"
requires-python = ">=3.11"
dependencies = [
    "agno",
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "anthropic",
    "supabase",
    "pydantic",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
